requirements.txt
# --- Core ---
logging
python-dotenv
PyYAML
openai

# --- Video Processing ---
moviepy==1.0.3
Pillow
imageio
imageio-ffmpeg
srt

# --- Web Server ---
flask
boto3
flask_cors


# --- Optional (but recommended) ---
requests
gunicorn
gtts